+++ src/AngularSPAWebAPI/app/services/auth.guard.ts
@@ -10,32 +10,38 @@ import { AuthenticationService } from './authentication.service';
  */
 @Injectable() export class AuthGuard implements CanActivate {
 
+    private signedIn: boolean;
+
     constructor(private authenticationService: AuthenticationService, private router: Router) { }
 
     public canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | boolean {
-        return this.authenticationService.isSignedIn().map((signedIn: boolean) => {
-            const url: string = state.url;
+        return this.authenticationService.isSignedIn()
+            .map((signedIn: boolean) => { this.signedIn = signedIn; })
+            .flatMap(() => this.authenticationService.userChanged()
+                .map(() => {
+                    const url: string = state.url;
 
-            if (signedIn) {
-                if (url !== "/dashboard") {
-                    return true;
-                } else {
-                    if (this.authenticationService.isInRole("administrator")) {
-                        return true;
-                    } else {
-                        this.router.navigate(['/home']);
-                        return false;
+                    if (this.signedIn) {
+                        if (url !== "/dashboard") {
+                            return true;
+                        } else {
+                            if (this.authenticationService.isInRole("administrator")) {
+                                return true;
+                            } else {
+                                this.router.navigate(['/home']);
+                                return false;
+                            }
+                        }
                     }
-                }
-            }
 
-            // Stores the attempted URL for redirecting.
-            this.authenticationService.redirectUrl = url;
+                    // Stores the attempted URL for redirecting.
+                    this.authenticationService.redirectUrl = url;
 
-            // Not signed in so redirects to signin page.
-            this.router.navigate(['/account/signin']);
-            return false;
-        });
+                    // Not signed in so redirects to signin page.
+                    this.router.navigate(['/account/signin']);
+                    return false;
+                })
+            );
     }
 
 }